--- src/components/home/navbar/navbar.module.css
@@ -1,32 +1,0 @@
-.container {
-  display: flex;
-  justify-content: flex-start;
-  align-items: center;
-  max-width: 100vw;
-  height: 100px;
-  background: black;
-
-  padding: 0 4em;
-}
-
-.logo {
-  width: 80px;
-  height: 80px;
-  border-radius: 360px;
-}
-
-.linkContainer {
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-  width: 100%;
-}
-
-.linkContainer > p {
-  color: white;
-  font-size: 18px;
-  font-weight: bold;
-  font-family: "Rubik Mono One", sans-serif;
-
-  margin: 0 2.5em;
-}
