+++ src/shared/navbar/logo.png
