+++ src/shared/navbar/navbar.module.css
@@ -1,0 +1,59 @@
+.container {
+  display: flex;
+  justify-content: flex-start;
+  align-items: center;
+  max-width: 100vw;
+  height: 60px;
+  background: black;
+
+  position: sticky;
+  top: 0;
+
+  padding: 0 4em;
+}
+
+.logo {
+  position: relative;
+  top: 10px;
+  width: 80px;
+  height: 80px;
+  border-radius: 360px;
+}
+
+.logo::before {
+  content: "";
+  position: absolute;
+  top: -10px;
+  width: 80px;
+  height: 70px;
+  background-color: rgba(143, 20, 20, 1);
+}
+
+.linkContainer {
+  display: flex;
+  justify-content: flex-end;
+  align-items: center;
+  height: 30%;
+  width: 100%;
+}
+
+.linkContainer > a {
+  color: white;
+  font-size: 18px;
+  font-weight: bold;
+  padding: 0 20px;
+
+  font-family: "Rubik", sans-serif;
+
+  margin: 0 0.5em;
+}
+
+.linkContainer > a > p {
+  padding: 7px 15px;
+}
+.current {
+  border-radius: 30px;
+
+  text-align: center;
+  background-color: rgba(143, 20, 20, 1);
+}
